CLI - client.conf - update path to default to /katello
This commit changes the default path used for client.conf to be /katello.
This will allow CLI to function with the new /katello prefix.

It should be noted, that I left the default for scheme as https; however,
at the time of this commit, production deployment will use https, but
development env will use http.