840531 - Fixes issue with inability to individually promote packages attached to a system template or changeset that have more than a single dash in the name.
This bug fix includes the removal of adding a package to a system
template by NVRE since this not only complicates the issue but is
inconsistent.  NVRE is not availabe for adding packages in the UI, and
is never used by any aspect of the system template.

Removing check for package existence on add_package call since
SystemTemplatePackage validation checks in the environment of the
template for existence already.