855267 - CLI add product label and id whereever product name is supported
This is an initial commit for 855267; however, there may be additional
changes as testing is completed.

This commit contains changes so that anywhere a 'product name' can
be used on the CLI the user may alternatively provide product
label or id.  This is necessary because product name is not unique
(i.e. multiple products may have the same name).  As a result,
the user should provide a unique identifier for the product
when using the CLI (e.g. id).

Without this change, the CLI behavior is essentially to always use
the 'first' product returned with the name specified, which can
result incorrect behaviors.

Note: we are leaving product name in place for backward compatability
for cases where there really is only 1 product w/ a given name.  That
said, CLI will now generate an error to the user if they use a name
and it returns multiple products with that name.  That way the user may
resubmit the command using product id (for example) and ensure they
are using the desired product.