859892 - system info contains "u"-s in "OS release" field
this was a unicode issue. the 'u's where popping up because python was
printing an entire hash instead of a string. the 'u' signifies it's
unicode string.

looking at consumer.rb for candlepin, you can find a 'release' method
that pulls the string out of that 'releaseVer' hash which is nicer for
the client to print.