855267 - CLI - pylint, template and changeset fixes
This commit contains a couple of minor changes raised by pylint;
however, those actually exposed additional changes needed to
support the product name/label/id for changesets and templates.

Essentially, it identified that the templates had not been updated
to support the product label and id.

Also, due to similarities between the code for changesets and templates
it found found some issues with changesets, such as ensuring
that name, label, id is treated as mutually exclusive.