pulpv2 - update to use 'finished' for task status state
Prepare for long rambling commit message... :)

This commit contains changes to use 'finished' vs
'success' for task status state.  The original
change from 'finished' to 'success' was actually due to
an inconsistency in the result from pulp task status and
pulp sync history.  At the time, the syncs transitioned
very quickly from being status to history; therefore,
the thought was that pulpv2 had also changed task status
to use 'success' as a final state.  Since it did not and
the task status is used for more than sync tasks, we are
reverting back to using 'finished'.