Spec change to deal with pylint issues
Sounds like katello cli has issues with pylint 1.0 which comes with f19.
Going to file a issue to fix the pylint suggested changes, but for now
sticking with the pylint the version that we use for f18 and rhel 6. We
should try to fix this in the future
(cherry picked from commit 851bea8feae3b0d15e597b5b171146215775adb5)